WIP - checkpoint after (hopefully) final restructuring/renaming.
