+++ src/types/instances.jl
@@ -1,5 +1,19 @@
 using DataStructures
 
+mutable struct Timestep{Offset, Duration, Final}
+	t::Int
+end
+
+mutable struct Clock
+	ts::Timestep
+
+	function Clock(offset::Int, final::Int, duration::Int)
+		self = new()
+		self.ts = Timestep{offset, duration, final}(1)
+		return self
+	end
+end
+
 abstract type ComponentInstanceData end
 
 # An instance of this type is passed to the run_timestep function of a
@@ -55,47 +69,3 @@ mutable struct ModelInstance
     offsets::Vector{Int}        # in order corresponding with components
     final_times::Vector{Int}
 end
-
-function _get_index_pos(names, propname, var_or_par)
-    index_pos = findfirst(names, propname)
-    index_pos == 0 && error("Unknown $var_or_par name $propname.")
-    return index_pos
-end
-
-
-#
-# Support for dot-overloading in run_timestep functions
-#
-
-# This is shared by parameter and variable get_property methods
-function _get_property_expr(obj::ComponentInstanceData, types, index_pos)
-    if types.parameters[index_pos] <: Ref
-        return :(obj.vals[$index_pos][])
-    else
-        return :(obj.vals[$index_pos])
-    end
-end
-
-@generated function get_property(p::ComponentInstanceParameters{NAMES,TYPES}, 
-                                 ::Val{PROPERTYNAME}) where {NAMES,TYPES,PROPERTYNAME}
-    index_pos = _get_index_pos(NAMES, PROPERTYNAME, "parameter")
-    return _get_property_expr(p, TYPES, index_pos)
-end
-
-@generated function get_property(v::ComponentInstanceVariables{NAMES,TYPES}, 
-                                 ::Val{PROPERTYNAME}) where {NAMES,TYPES,PROPERTYNAME}
-    index_pos = _get_index_pos(NAMES, PROPERTYNAME, "variable")
-    return _get_property_expr(v, TYPES, index_pos)
-end
-
-@generated function set_property!(v::ComponentInstanceVariables{NAMES,TYPES}, 
-                                  ::Val{PROPERTYNAME}, value) where {NAMES,TYPES,PROPERTYNAME}
-    index_pos = _get_index_pos(NAMES, PROPERTYNAME, "variable")
-
-    if TYPES.parameters[index_pos] <: Ref
-        return :(v.vals[$index_pos][] = value)
-    else
-        error("You cannot override indexed variable $PROPERTYNAME.")
-    end
-end
-
